Inline single-use step traits and remove redundant null tokens
