--- src/Gateway/Gemini/Concerns/HandlesTextSteps.php
@@ -1,83 +1,0 @@
-<?php
-
-namespace Laravel\Ai\Gateway\Gemini\Concerns;
-
-use Generator;
-use Laravel\Ai\Contracts\Providers\TextProvider;
-use Laravel\Ai\Gateway\StepContext;
-use Laravel\Ai\Gateway\StepResponse;
-use Laravel\Ai\Gateway\TextGenerationOptions;
-
-trait HandlesTextSteps
-{
-    /**
-     * Generate text for a single generateContent step.
-     */
-    public function generateTextStep(
-        TextProvider $provider,
-        string $model,
-        ?string $instructions,
-        array $messages,
-        array $tools,
-        ?array $schema,
-        ?TextGenerationOptions $options,
-        ?int $timeout,
-        StepContext $stepContext,
-    ): StepResponse {
-        $body = $this->buildStepBody($provider, $model, $instructions, $messages, $tools, $schema, $options, $stepContext);
-
-        $response = $this->withErrorHandling(
-            $provider->name(),
-            fn () => $this->client($provider, $timeout)->post("models/{$model}:generateContent", $body),
-        );
-
-        $data = $response->json();
-
-        $this->validateTextResponse($data);
-
-        return $this->parseTextResponse($data, $provider, $model, filled($schema));
-    }
-
-    /**
-     * Stream text for a single streamGenerateContent step.
-     */
-    public function generateStreamStep(
-        string $invocationId,
-        TextProvider $provider,
-        string $model,
-        ?string $instructions,
-        array $messages,
-        array $tools,
-        ?array $schema,
-        ?TextGenerationOptions $options,
-        ?int $timeout,
-        StepContext $stepContext,
-    ): Generator {
-        $body = $this->buildStepBody($provider, $model, $instructions, $messages, $tools, $schema, $options, $stepContext);
-
-        $response = $this->withErrorHandling(
-            $provider->name(),
-            fn () => $this->client($provider, $timeout)
-                ->withOptions(['stream' => true])
-                ->post("models/{$model}:streamGenerateContent?alt=sse", $body),
-        );
-
-        return yield from $this->processTextStream($invocationId, $provider, $model, $response->getBody());
-    }
-
-    /**
-     * Build the request body for the current text generation step.
-     */
-    protected function buildStepBody(
-        TextProvider $provider,
-        string $model,
-        ?string $instructions,
-        array $messages,
-        array $tools,
-        ?array $schema,
-        ?TextGenerationOptions $options,
-        StepContext $stepContext,
-    ): array {
-        return $this->buildTextRequestBody($provider, $instructions, $messages, $tools, $schema, $options);
-    }
-}
